requirements.txt
flask
flask-restplus
redis
gutter-python3